memory_leak_x2 doesn't have to be long anymore

+++ pandaserver/taskbuffer/OraDBProxy.py
@@ -2259,7 +2259,7 @@ class DBProxy:
                     tmpM = re.search('chi2=(-?\d+\.*\d+)', param[key])
                     if tmpM is not None:
                         # keep measurement under 11 digits because of DB declaration
-                        memory_leak_x2 = min(long(float(tmpM.group(1))), long(10**11-1))
+                        memory_leak_x2 = min(float(tmpM.group(1)), 10**11-1)
                         tmpKey = 'memory_leak_x2'
                         sql1 += ',{0}=:{0}'.format(tmpKey)
                         varMap[':{0}'.format(tmpKey)] = memory_leak_x2
